+++ lib/investigationCollection/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./dnsProvider";
+export * from "./productionProviderAdapters";
 export * from "./orchestrator";